Moved some functions in model to util folder
